fix: wrong renderer used due to memoization
When a document is printed multiple times, memoization is reused, as intended.
However, these printing could be done with different renderers.
Prior this commit, the memoization also remembers the renderer that was used.
Therefore, subsequent printing could use a stale, memoized renderer.
This commit fixes the issue by threading the renderer through the layout
function, so that it becomes "stateless" and can be memoized with no
stale state saving.

Fixes #2